// graphql-types.ts
export type Maybe<T> = T | null
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /**
   * A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any
}
// ...
export type Dates = Node & {
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  date?: Maybe<Scalars['Date']>
  totalItems?: Maybe<Scalars['Int']>
  totalEvents?: Maybe<Scalars['Int']>
  totalGames?: Maybe<Scalars['Int']>
  totalGamesInProgress?: Maybe<Scalars['Int']>
  games?: Maybe<Array<Maybe<DatesGames>>>
}
// ...
export type DatesGames = {
  gamePk?: Maybe<Scalars['Int']>
  link?: Maybe<Scalars['String']>
  gameType?: Maybe<Scalars['String']>
  season?: Maybe<Scalars['Date']>
  gameDate?: Maybe<Scalars['Date']>
  status?: Maybe<DatesGamesStatus>
  teams?: Maybe<DatesGamesTeams>
  venue?: Maybe<DatesGamesVenue>
  content?: Maybe<DatesGamesContent>
  gameNumber?: Maybe<Scalars['Int']>
  publicFacing?: Maybe<Scalars['Boolean']>
  doubleHeader?: Maybe<Scalars['String']>
  gamedayType?: Maybe<Scalars['String']>
  tiebreaker?: Maybe<Scalars['String']>
  calendarEventID?: Maybe<Scalars['String']>
  seasonDisplay?: Maybe<Scalars['Date']>
  dayNight?: Maybe<Scalars['String']>
  scheduledInnings?: Maybe<Scalars['Int']>
  inningBreakLength?: Maybe<Scalars['Int']>
  gamesInSeries?: Maybe<Scalars['Int']>
  seriesGameNumber?: Maybe<Scalars['Int']>
  seriesDescription?: Maybe<Scalars['String']>
  recordSource?: Maybe<Scalars['String']>
  ifNecessary?: Maybe<Scalars['String']>
  ifNecessaryDescription?: Maybe<Scalars['String']>
  description?: Maybe<Scalars['String']>
}
// ...
export type DatesGamesContent = {
  link?: Maybe<Scalars['String']>
}
// ...
export type DatesGamesStatus = {
  abstractGameState?: Maybe<Scalars['String']>
  codedGameState?: Maybe<Scalars['String']>
  detailedState?: Maybe<Scalars['String']>
  statusCode?: Maybe<Scalars['String']>
  abstractGameCode?: Maybe<Scalars['String']>
}

export type DatesGamesStatusFilterInput = {
  abstractGameState?: Maybe<StringQueryOperatorInput>
  codedGameState?: Maybe<StringQueryOperatorInput>
  detailedState?: Maybe<StringQueryOperatorInput>
  statusCode?: Maybe<StringQueryOperatorInput>
  abstractGameCode?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesTeams = {
  away?: Maybe<DatesGamesTeamsAway>
  home?: Maybe<DatesGamesTeamsHome>
}
// ...
export type DatesGamesTeamsAwayTeam = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
  season?: Maybe<Scalars['Int']>
  venue?: Maybe<DatesGamesTeamsAwayTeamVenue>
  teamCode?: Maybe<Scalars['String']>
  fileCode?: Maybe<Scalars['String']>
  abbreviation?: Maybe<Scalars['String']>
  teamName?: Maybe<Scalars['String']>
  locationName?: Maybe<Scalars['String']>
  firstYearOfPlay?: Maybe<Scalars['Date']>
  league?: Maybe<DatesGamesTeamsAwayTeamLeague>
  division?: Maybe<DatesGamesTeamsAwayTeamDivision>
  sport?: Maybe<DatesGamesTeamsAwayTeamSport>
  shortName?: Maybe<Scalars['String']>
  springLeague?: Maybe<DatesGamesTeamsAwayTeamSpringLeague>
  allStarStatus?: Maybe<Scalars['String']>
  active?: Maybe<Scalars['Boolean']>
}

export type DatesGamesTeamsAwayTeamFirstYearOfPlayArgs = {
  formatString?: Maybe<Scalars['String']>
  fromNow?: Maybe<Scalars['Boolean']>
  difference?: Maybe<Scalars['String']>
  locale?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsAwayTeamDivision = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
}
// ...
export type DatesGamesTeamsAwayTeamLeague = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsAwayTeamLeagueFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesTeamsAwayTeamSport = {
  alternative_id?: Maybe<Scalars['Int']>
  link?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsAwayTeamSportFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesTeamsAwayTeamSpringLeague = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
  abbreviation?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsAwayTeamSpringLeagueFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
  abbreviation?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesTeamsAwayTeamVenue = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
}
// ...
export type DatesGamesTeamsHomeTeam = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
  season?: Maybe<Scalars['Int']>
  venue?: Maybe<DatesGamesTeamsHomeTeamVenue>
  teamCode?: Maybe<Scalars['String']>
  fileCode?: Maybe<Scalars['String']>
  abbreviation?: Maybe<Scalars['String']>
  teamName?: Maybe<Scalars['String']>
  locationName?: Maybe<Scalars['String']>
  firstYearOfPlay?: Maybe<Scalars['Date']>
  league?: Maybe<DatesGamesTeamsHomeTeamLeague>
  division?: Maybe<DatesGamesTeamsHomeTeamDivision>
  sport?: Maybe<DatesGamesTeamsHomeTeamSport>
  shortName?: Maybe<Scalars['String']>
  springLeague?: Maybe<DatesGamesTeamsHomeTeamSpringLeague>
  allStarStatus?: Maybe<Scalars['String']>
  active?: Maybe<Scalars['Boolean']>
}

export type DatesGamesTeamsHomeTeamFirstYearOfPlayArgs = {
  formatString?: Maybe<Scalars['String']>
  fromNow?: Maybe<Scalars['Boolean']>
  difference?: Maybe<Scalars['String']>
  locale?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsHomeTeamDivision = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
}
// ...
export type DatesGamesTeamsHomeTeamLeague = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsHomeTeamLeagueFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesTeamsHomeTeamSport = {
  alternative_id?: Maybe<Scalars['Int']>
  link?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsHomeTeamSportFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesTeamsHomeTeamSpringLeague = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
  abbreviation?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsHomeTeamSpringLeagueFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
  abbreviation?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesTeamsHomeTeamVenue = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
}

export type DatesGamesTeamsHomeTeamVenueFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
}

export type DatesGamesVenue = {
  alternative_id?: Maybe<Scalars['Int']>
  name?: Maybe<Scalars['String']>
  link?: Maybe<Scalars['String']>
  location?: Maybe<DatesGamesVenueLocation>
}

export type DatesGamesVenueFilterInput = {
  alternative_id?: Maybe<IntQueryOperatorInput>
  name?: Maybe<StringQueryOperatorInput>
  link?: Maybe<StringQueryOperatorInput>
  location?: Maybe<DatesGamesVenueLocationFilterInput>
}

export type DatesGamesVenueLocation = {
  city?: Maybe<Scalars['String']>
  state?: Maybe<Scalars['String']>
  stateAbbrev?: Maybe<Scalars['String']>
  defaultCoordinates?: Maybe<DatesGamesVenueLocationDefaultCoordinates>
}

export type DatesGamesVenueLocationDefaultCoordinates = {
  latitude?: Maybe<Scalars['Float']>
  longitude?: Maybe<Scalars['Float']>
}
// ...
export type Internal = {
  content?: Maybe<Scalars['String']>
  contentDigest: Scalars['String']
  description?: Maybe<Scalars['String']>
  fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>
  ignoreType?: Maybe<Scalars['Boolean']>
  mediaType?: Maybe<Scalars['String']>
  owner: Scalars['String']
  type: Scalars['String']
}
// ...
export type Node = {
  id: Scalars['ID']
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type GetScheduleQuery = {
  allDates: {
    nodes: Array<
      Pick<Dates, 'id' | 'date'> & {
        games: Maybe<
          Array<
            Maybe<
              Pick<DatesGames, 'gamePk' | 'gameDate'> & {
                venue: Maybe<
                  Pick<DatesGamesVenue, 'name'> & {
                    location: Maybe<
                      Pick<DatesGamesVenueLocation, 'city' | 'state'> & {
                        defaultCoordinates: Maybe<
                          Pick<
                            DatesGamesVenueLocationDefaultCoordinates,
                            'latitude' | 'longitude'
                          >
                        >
                      }
                    >
                  }
                >
                teams: Maybe<{
                  home: Maybe<{
                    team: Maybe<
                      Pick<DatesGamesTeamsHomeTeam, 'name' | 'abbreviation'>
                    >
                  }>
                  away: Maybe<{
                    team: Maybe<
                      Pick<DatesGamesTeamsAwayTeam, 'name' | 'abbreviation'>
                    >
                  }>
                }>
              }
            >
          >
        >
      }
    >
  }
}